feat: preparar backend para Sprint 2
Se reorganizó y actualizó el backend para adaptarlo a la nueva estructura de base de datos, ajustando entidades, repositorios, servicios y controladores. Se corrigieron tipos de datos, relaciones y validaciones, y se incorporaron los módulos de solicitudes, plantillas, versiones, campos, secciones, elementos y certificados generados, junto con el registro de logs.

Se implementó y probó el flujo de generación de certificados en PDF a partir de plantillas, incluyendo posiciones, tamaños y alineación de elementos. También se realizaron pruebas de integración entre el backend y la base de datos, verificando la compilación, el inicio de la aplicación y el flujo completo desde la solicitud hasta la generación y almacenamiento del certificado, dejando una base funcional para continuar el desarrollo durante el Sprint 2.

+++ backend/src/main/java/com/certificados/app/controller/ProgramaController.java
@@ -26,7 +26,7 @@ public class ProgramaController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ProgramaDTO> buscarPorId(@PathVariable Long id) {
+    public ResponseEntity<ProgramaDTO> buscarPorId(@PathVariable Integer id) {
         return ResponseEntity.ok(programaService.buscarPorId(id));
     }
